# backend/docs/swagger.yaml
basePath: /
definitions:
...
  ai-models-backend_internal_models.CreateFeedCommentRequest:
    properties:
      content:
        description: 评论内容，最多500字符
        maxLength: 500
        type: string
      replyTo:
        description: 回复的用户名
        maxLength: 100
        type: string
    required:
    - content
    type: object
  ai-models-backend_internal_models.CreateFeedPostRequest:
    properties:
      content:
        description: 最多2000字符
        maxLength: 2000
        type: string
      imageURL:
        description: 图片URL
        type: string
    type: object
...
  ai-models-backend_internal_models.FeedComment:
    properties:
      avatar:
        description: 冗余头像
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      likeCount:
        type: integer
      postID:
        type: string
      replyTo:
        description: 回复的用户名
        type: string
      updatedAt:
        type: string
      userID:
        type: string
      userProfileVersion:
        description: 用户信息版本号
        type: integer
      username:
        description: 冗余字段
        type: string
    type: object
  ai-models-backend_internal_models.FeedCommentResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/ai-models-backend_internal_models.FeedComment'
        type: array
      hasMore:
        type: boolean
      nextCursor:
        type: string
      total:
        type: integer
    type: object
  ai-models-backend_internal_models.FeedPost:
    properties:
      avatar:
        description: 冗余头像
        type: string
      commentCount:
        type: integer
      content:
        description: 文字内容（可选）
        type: string
      createdAt:
        type: string
      id:
        type: string
      imageURL:
        description: 图片URL（可选）
        type: string
      likeCount:
        type: integer
      status:
        description: 用户状态emoji
        type: string
      updatedAt:
        type: string
      userID:
        type: string
      userProfileVersion:
        description: 用户信息版本号
        type: integer
      username:
        description: 冗余字段
        type: string
    type: object
  ai-models-backend_internal_models.FeedPostResponse:
    properties:
      hasMore:
        type: boolean
      nextCursor:
        type: string
      posts:
        items:
          $ref: '#/definitions/ai-models-backend_internal_models.FeedPost'
        type: array
    type: object
...
  ai-models-backend_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: AI Models Backend API documentation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: AI Models API
  version: "1.0"
paths:
...
  /api/feed/comments/{comment_id}/like:
    post:
      consumes:
      - application/json
      description: 点赞或取消点赞评论，需要登录
      parameters:
      - description: 评论ID
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
      summary: 切换评论点赞状态
      tags:
      - Feed
  /api/feed/posts:
    get:
      consumes:
      - application/json
      description: 支持多种排序方式和cursor分页
      parameters:
      - default: time
        description: 排序类型
        enum:
        - time
        - like
        - comment
        in: query
        name: sort
        type: string
      - description: cursor分页的after_id
        in: query
        name: after_id
        type: string
      - default: 20
        description: 每页数量
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai-models-backend_internal_models.FeedPostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
      summary: 获取信息流帖子列表
      tags:
      - Feed
    post:
      consumes:
      - application/json
      description: 创建新的信息流帖子，需要登录
      parameters:
      - description: 帖子内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai-models-backend_internal_models.CreateFeedPostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai-models-backend_internal_models.FeedPost'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
      summary: 创建信息流帖子
      tags:
      - Feed
  /api/feed/posts/{post_id}:
    get:
      consumes:
      - application/json
      description: 获取指定帖子的详细信息
      parameters:
      - description: 帖子ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai-models-backend_internal_models.FeedPost'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
      summary: 获取帖子详情
      tags:
      - Feed
  /api/feed/posts/{post_id}/comments:
    get:
      consumes:
      - application/json
      description: 获取指定帖子的评论列表，支持cursor分页
      parameters:
      - description: 帖子ID
        in: path
        name: post_id
        required: true
        type: string
      - description: cursor分页的after_id
        in: query
        name: after_id
        type: string
      - default: 20
        description: 每页数量
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai-models-backend_internal_models.FeedCommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
      summary: 获取帖子评论列表
      tags:
      - Feed
    post:
      consumes:
      - application/json
      description: 为指定帖子创建评论，需要登录
      parameters:
      - description: 帖子ID
        in: path
        name: post_id
        required: true
        type: string
      - description: 评论内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai-models-backend_internal_models.CreateFeedCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai-models-backend_internal_models.FeedComment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
      summary: 创建帖子评论
      tags:
      - Feed
  /api/feed/posts/{post_id}/like:
    post:
      consumes:
      - application/json
      description: 点赞或取消点赞帖子，需要登录
      parameters:
      - description: 帖子ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
      summary: 切换帖子点赞状态
      tags:
      - Feed
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"